Style and naming fixes

+++ demo/ffmpeg-to-shaka-packager/dashTestPattern.sh
@@ -1,15 +1,15 @@
 #!/bin/bash
 
-port=8080
+PORT=8080
 # create a unique directory
-dir=$(date '+%m-%d-%y-%T')
-vid='TestPattern'
+OUTPUT_DIR=$(date '+%m-%d-%y-%T')
+OUTPUT_SEG_NAME='test_pattern_live_video'
 
-rm pipe0
+[ -e pipe0 ] && rm pipe0
 mkfifo pipe0
 
 # encode with FFMPEG
-x264enc='libx264 -tune zerolatency -profile:v high -preset ultrafast -bf 0 -refs 3 -sc_threshold 0'
+X264_ENC='libx264 -tune zerolatency -profile:v high -preset ultrafast -bf 0 -refs 3 -sc_threshold 0'
 
 ffmpeg \
     -hide_banner \
@@ -18,7 +18,7 @@ ffmpeg \
     -i "testsrc2=size=1920x1080:rate=30" \
     -pix_fmt yuv420p \
     -map 0:v \
-    -c:v ${x264enc} \
+    -c:v ${X264_ENC} \
     -g 150 \
     -keyint_min 150 \
     -b:v 4000k \
@@ -29,8 +29,9 @@ ffmpeg \
 # package as DASH
 packager \
    --io_block_size 65536 \
-   in=pipe0,stream=video,init_segment='http://0.0.0.0:8080/'${dir}'/'${vid}'_live_video_init.m4s',segment_template='http://0.0.0.0:'${port}'/'${dir}'/'${vid}'_live_video_$Number$.m4s' \
+   in=pipe0,stream=video,init_segment='http://0.0.0.0:'${PORT}'/'${OUTPUT_DIR}'/'${OUTPUT_SEG_NAME}'_init.m4s',segment_template='http://0.0.0.0:'${PORT}'/'${OUTPUT_DIR}'/'${OUTPUT_SEG_NAME}'_$Number$.m4s' \
    --segment_duration 5 \
-   --mpd_output "http://0.0.0.0:${port}/${dir}/manifest.mpd"
+   --mpd_output "http://0.0.0.0:${PORT}/${OUTPUT_DIR}/manifest.mpd" \
+    >& log.log
    
 rm pipe0
